spin some stuff out to util for readability
