+++ index.js
@@ -5,10 +5,9 @@ import { Filters } from './modules/filters.js'
 import { Language } from './modules/language.js'
 import { Spacing } from './modules/spacing.js'
 import { Theme } from './modules/theme.js'
+import * as util from './modules/util.js'
 
 const fontsize = new Fontsize()
-
-let fontData
 
 window.CMeditor = CodeMirror.fromTextArea(document.getElementById('code'), {
   lineNumbers: true,
@@ -17,108 +16,6 @@ window.CMeditor = CodeMirror.fromTextArea(document.getElementById('code'), {
   theme: 'pastel-on-dark',
   lineWrapping: true
 })
-
-function isVisible (el) {
-  const container = document.querySelector('section.select-list').getBoundingClientRect()
-  const target = el.getBoundingClientRect()
-
-  return target.bottom > container.top && target.top < container.bottom
-}
-
-/**
- * Get the first visible (non-filtered) entry's alias from the rendered list
- */
-function getFirstEntryAlias () {
-  const first = document.querySelector('#select-font .entry:not(.filtered-out):not(.group-child)')
-  return first ? first.getAttribute('data-alias') : null
-}
-
-/**
- * Get the font from the # or the top-most entry
- */
-function getFont () {
-  let font = window.location.hash.substring(1)
-
-  if (!font) {
-    font = getFirstEntryAlias()
-  }
-
-  return font
-}
-
-function setDetails (data) {
-  const box = document.querySelector('.info-wrapper')
-  box.querySelector('h2').setAttribute('data-license', data.license ?? '')
-  box.querySelector('a').href = data.website
-  box.querySelector('h2 a').textContent = data.name
-  box.querySelector('p.info').textContent = data.description ?? ''
-  if (typeof data.variants === 'string') {
-    box.querySelector('p.variants').textContent = data.variants
-  } else if (data.variants && data.variants.length > 1) {
-    box.querySelector('p.variants').textContent = data.variants.join(', ')
-  } else {
-    box.querySelector('p.variants').textContent = ''
-  }
-}
-
-// ProgrammingFonts font selector
-function selectFont () {
-  const codeMirror = document.querySelector('.CodeMirror')
-  const font = getFont()
-
-  if (typeof fontData === 'undefined' || typeof fontData[font] === 'undefined') {
-    return
-  }
-
-  setDetails(fontData[font])
-  codeMirror.setAttribute('data-font', font)
-
-  if (fontData[font].rendering === 'bitmap') {
-    codeMirror.classList.add('no-smooth')
-    if (fontData[font]['bitmap size']) {
-      fontsize.forceSize(fontData[font]['bitmap size'])
-    }
-  } else {
-    codeMirror.classList.remove('no-smooth')
-    fontsize.reset()
-  }
-
-  if (font === 'input') {
-    // because Input Mono is loaded via external @font-face file
-    codeMirror.style.fontFamily = 'Input Mono, monospace'
-    codeMirror.querySelectorAll('pre, textarea').forEach((element) => {
-      element.style.fontFamily = 'Input Mono, monospace'
-    })
-  } else {
-    codeMirror.style.fontFamily = `${font}, monospace`
-    codeMirror.querySelectorAll('pre, textarea').forEach((element) => {
-      element.style.fontFamily = `${font}, monospace`
-    })
-  }
-
-  document.querySelectorAll('#select-font [data-alias]').forEach((element) => {
-    element.classList.remove('active')
-  })
-
-  const activeEntry = document.querySelector(`#select-font [data-alias='${font}']`)
-  if (activeEntry) {
-    if (activeEntry.classList.contains('group-child')) {
-      const groupAlias = activeEntry.getAttribute('data-group')
-      const primary = document.querySelector(`#select-font [data-alias='${groupAlias}']`)
-      if (primary && !primary.classList.contains('group-open')) {
-        window.toggleGroup(groupAlias)
-      }
-    }
-    activeEntry.classList.add('active')
-    if (!isVisible(activeEntry)) {
-      activeEntry.scrollIntoView({
-        block: 'center',
-        inline: 'nearest',
-        behavior: 'smooth'
-      })
-    }
-  }
-}
 
 const chevronDownIcon = '<svg aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16" fill="currentColor"><path d="M12.78 5.22a.749.749 0 0 1 0 1.06l-4.25 4.25a.749.749 0 0 1-1.06 0L3.22 6.28a.749.749 0 1 1 1.06-1.06L8 8.939l3.72-3.719a.749.749 0 0 1 1.06 0Z"/></svg>'
 
@@ -250,7 +147,7 @@ function renderSelectList (selectFirst = false) {
     }).then((data) => {
       const fonts = []
 
-      fontData = data
+      window.fontData = data
 
       Object.keys(data).forEach((key) => {
         const v = data[key]
@@ -262,12 +159,12 @@ function renderSelectList (selectFirst = false) {
       new Filters(data, () => {renderSelectList(true)}).init()
 
       if (selectFirst) {
-        const first = getFirstEntryAlias()
+        const first = util.getFirstEntryAlias()
         if (first) {
           window.location.hash = first
         }
       }
-      selectFont()
+      util.selectFont()
     })
 }
 
@@ -312,33 +209,8 @@ function walk (direction) {
   }
 }
 
-function expandGroup () {
-  const activeEntry = document.querySelector('.entry.active')
-  if (!activeEntry || activeEntry.classList.contains('group-child')) {
-    return
-  }
-  if (activeEntry.querySelector('.group-toggle') && !activeEntry.classList.contains('group-open')) {
-    window.toggleGroup(activeEntry.getAttribute('data-alias'))
-  }
-}
-
-function collapseGroup () {
-  const activeEntry = document.querySelector('.entry.active')
-  if (!activeEntry) {
-    return
-  }
-  if (activeEntry.classList.contains('group-child')) {
-    const primary = document.querySelector(`#select-font [data-alias='${activeEntry.getAttribute('data-group')}']`)
-    if (primary) {
-      primary.querySelector('a').click()
-    }
-  } else if (activeEntry.classList.contains('group-open')) {
-    window.toggleGroup(activeEntry.getAttribute('data-alias'))
-  }
-}
-
 window.onhashchange = () => {
-  selectFont()
+  util.selectFont()
 }
 
 window.addEventListener('DOMContentLoaded', () => {
@@ -363,11 +235,11 @@ window.addEventListener('DOMContentLoaded', () => {
     } else if (event.key === 'ArrowRight') {
       event.preventDefault()
       event.stopPropagation()
-      expandGroup()
+      util.expandGroup()
     } else if (event.key === 'ArrowLeft') {
       event.preventDefault()
       event.stopPropagation()
-      collapseGroup()
+      util.collapseGroup()
     }
   }
